requirements.txt
altair==5.4.1
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
branca==0.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
findspark==2.0.1
folium==0.19.4
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
narwhals==1.8.1
nest-asyncio==1.6.0
numpy==2.1.1
packaging==23.2
pandas==2.2.2
parso==0.8.3
pillow==10.4.0
platformdirs==4.2.0
plotly==5.24.1
prompt-toolkit==3.0.43
protobuf==5.28.2
psutil==5.9.8
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.17.2
pyspark==3.5.1
python-dateutil==2.8.2
pytz==2024.2
pywin32==306
pyzmq==25.1.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
wcwidth==0.2.13
xyzservices==2025.1.0
